pyproject.toml
[project]
name = "ocr-gauntlet"
version = "0.1.0"
description = "5 tiers of OCR compared: from Tesseract to Gemini 3 Flash"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "Slava Dubrov" }]

dependencies = [
    "Pillow>=10.0",
    "numpy>=1.24",
    "pandas>=2.0",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "tqdm>=4.65",
    "rich>=13.0",
    "editdistance>=0.8",
    "jiwer>=3.0",           # WER computation
    "datasets>=2.14",       # HuggingFace datasets for sample downloads
]

[project.optional-dependencies]
tesseract = ["pytesseract>=0.3.10"]
docling = ["docling>=2.70"]
gemini = ["google-generativeai>=0.8.0"]
mistral = ["mistralai>=1.0.0"]
dots-ocr = ["openai>=1.0"]    # dots.ocr client; vLLM server installed separately (needs GPU + Python <=3.13)
notebooks = ["jupyter>=1.0", "ipywidgets>=8.0"]
all = ["ocr-gauntlet[tesseract,docling,gemini,mistral,dots-ocr,notebooks]"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ocr_gauntlet"]

[dependency-groups]
dev = ["pytest>=7.0", "ruff>=0.4"]
